+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add require('@openveo/api').watcher to listen to files system changes
 - ovWatch script has been added to execute an NPM script of files changes and can be launched using npx (see scripts/watch.js) for usage
 - ovRemove script has been added to remove a file system resource and can be launched using npx (see scripts/remove.js) for usage
+- require('@openveo/api').imageProcessor.generateSpriteFreely has been added to be able to generate a sprite without resizing the images and simply ordering images by their height with all images of the same height on the same line of the sprites
 
 ## BUG FIXES